Made conversions into a protocol
The separation between apps made the once-simple conversions code into
a pretty gross mess of map-based pattern matching, with the to_lsp
bits stripped out. Time for a protocol to solve the problem of adding
implementations across apps. The result is that the protocol app can
add an implementation for the ranged protocol for any of the types it
defines, and all can be converted to and from native utf8 ranges easily.