Don't always clear diagnostics on save
Prior, we cleared the project's diagnostics when a file in the project
was saved. In reality, we should only clear a file's diagnostics on
project compilation finishing if the file on disk is the same as our
internal representation. This means, only clear it if the file is both
open and dirty.