Improved handling of parse-time errors
This has been annoying me for years. If there was a parse error, the
LSP used to just sit there, dead and not tell you what happened, nor
would it highlight the problem. I would have to format the code to see
the error.

This commit changes that behavior so that not only are parse errors
emitted by the language server, but they're parsed and their sources
(the helpful tips that point to the source of the error, like "the (
on line 3 is missing a )") are highlighted as well as where the
diagnostic points to.