Lexical coludn't format some of lexical's projects
I'm pretty sure these fixes are specific to lexical, but getting the
formatter.exs was failing in remote control because elixir said that
proto wasn't defined, but that's odd, since the .formatter.exs in
remote control didn't import that dep. Setting the dep to optional
fixed it.

I also found a case where formatting mix.exs would fail because a
diagnostic context wasn't found, so I defaulted the position to 0