Improved module purging
Prior, we relied on heuristics whether to purge a module, which
resulted in some modules not existing (they had no content, so it
didn't matter much for our purposes). However, the heuristic would
miss if you changed a module's name, and it would exhibit the exact
same problems we were having before, defining submodules as you
typed.

This fix introduces a mapping of module -> file path, so when we
compile a file, we can tell which modules were previously defined in
it. If a module previously existed in this file, but doesn't now, we
purge it, as it's been deleted from this file. This elimitates the
error of defining "partial" modules.

Fixes #32