Project node monitoring
While debuggin the issue of diagnostics not being reset when a file is
saved / compiled, I noticed that the cause was that the build process
died and couldn't complete until the build directory was deleted.

Due to the way things were supervised, the build node would remain up
and other stuff would work, but project builds wouldn't even be
attempted because the remote_control app on the project node was
dead.

The solution presented here is to monitor the remote control
supervisor in a process on the manager node and respond to a failure
of the supervisor by deleting the build directory and restarting the
node from scratch. This also moves supervison away from the dispatch
process (which was a bit strange) and into its own process that's just
responsible for the health of the remote node.