Added a namespace task
When editing the remote_control app while lexical is running as your
language server, it's easy to change a file and change lexical's
behavior. To prevent this, the task `mix namespace` will read all the
beam files, strip out any mentions of `Elixir.Lexical` in atoms and
replace that with `Elixir.LXRelease`.

There are a copule caveats:
1. The mod: keys in the mix.exs files in server and remote_control need to
   know about namespacing
2. I'm going to bet this wreaks havoc with test files.
3. My lsp doesn't connect to the server when it's namespaced, so
   there's more work to do