Converted range-able response types
I was noticing a lot of manual conversions in the code, and that kind
of stinks. To help with this, I'm going to change how things work in
the project, but this requires some additions to the protocol.

The overarching idea is to push conversions into the protocol, so that
on the way in, we convert from LSP structures to native and on the way
out we convert from native to LSP structures. Presently, there was no
conversion on the way out (there is conversion on the way in), so this
PR adds conversion when we write data.

The next step will be to change our native structures to be
one-based (thanks, Elixir, for not making that configurable) so all we
need to deal with in the language server are our data structures.