Simplified building requests and notifications
The protocol acutally defines params types for notifications and
responses,  and we were recapitulating them in their definitions. This
PR adds a bunch of  the auto-generated types and allows
notification/request  definitions to reference them. Then the types
are  consulted to build the request or notification's parameters.

The result is much simpler definitions and less manual work when we
need  to create new ones.