Dialyzer analysis (#120)
* Dialyzer analysis

To prepare for adding dialyzer diagnostics to the project, it would be
nice if we ran dialyzer clean. This PR fixes a number of issues that
dialyzer found, and actually fixes a couple of bugs that were going
unreported.
In addition, it enables dialyzer analysis in our github actions
workflow, so we can keep the project clean.

I had to ignore a couple of things in the configuration code, as it
would be impossible for it to dialyze clean as long as we don't have
a way to enable dialyzer.