Fixed object literal module names
Object literals are defined inside the module that uses them, and then
referred to via a local module name. This causes an issue with the
decoding logic, because these module names aren't fully qualified, and
decoding fails.

The fix here puts the full module name in calls to `deftype` and
friends so the modules can be resolved during decoding