Improved function capture completions
Function captures used to only complete `&Mod.fun/arity`, which is
_ok_, but sometimes you want to do `&Mod.fun(&1)`. This PR adds two
completion options for each function, one for the "complete" capture,
and another that lets you add arguments.

I'm not sure if I like the UX on this yet, but I do like that I can
pick between the two options. I'm open to suggestions about how to
improve the UX.