Refactor: Make callable completions consistent
We spent a bunch of time working on function completions in a variety
of cases, but macro completions still were quite primitive, only
completing with the name and arity, which is almost never what you
want.

This refactors the logic for building these completions into a
`Callable` module so both macros and functions can share their
behavior.

Fixes #136