Addressed deadlocks in Document Store
I was getting sporadic deadlocks in the document store since I moved
to lsp-mode (it sends a *lot* more didChange updates than eglot), so
I'm trying moving writes out of the genserver to see if that helps (it
seems to).

I also noticed some troubling messages about not being able to apply
didChange events, but that was due to an incorrect log message in the
server state, which is also fixed here.