Added file watching
If the server supports it, we will now receive watched file
notifications for files that have been added / updated or deleted in
the project directories.

This will help determining whether we should recompile the whole
project if a dependency has changed, or if we should remove a module
from our compiled module cache.