Added fragment capabilities to document
While talking with Jose Valim about how he could help lexical, we
discussed giving a document fragment to a function in
Code.Fragment. It seemed to me at the time that one of the functions
already there might fit the bill, so I coded up the ability to request
a fragment of a document using positions.

Unfortunately, the function in question won't work for us, but I felt
that document fragments might be helpful in the future, so here's a PR.