Fixed code actions / improved code mod api
Code actions were kind of broken, they were unable to process all
elements of an elixir document. They would fail if one of the clauses
in a case statement had an unused variable, for example. This was due
to the old code processing one line at a time.

The new method uses Sourceror Zippers, which can focus on a line of
code and process it. Even better, we can use Sourceror patches to
generate a list of patches, which can then be turned directly into
edits with no string processing at all. This new way of dealing with
code is a lot easier than it was previously, and i'm much happier with
the API.