New project structure / beginning of plugins
Making plugins necessitated a new project structure, as we're going to
need to publish a separate library for plugin authors so they can use
them. We'll also need a bunch of the modules they'll use, so we need
the least stressful way of depending on the libraries we publish.

Creating a separate repo would make our lives very difficult, so I
added a projects directory at the top level, and the published
libraries go in there. We utilize path dependencies for our internal
libraries.

I also added the beginnings of a plugin architecture, but I wanted to
get this change out before it gets too big. Plugins are quite
complicated, and we're going to need to do a bit of refactoring when
we add one. That refactoring, plus all of this moving would make the
PR very difficult to review.

This PR also includes a new genserver dispatcher so that the compile
tracer isn't slowed down by extraneous work.