Made best-effort to find formatter options
If we can't find the formatter, we were running the formatter with
default options, which would be annoying if you've configured
the formatter, and you'd just get the defaults. This could be
annoying. This change makes a best effort to find the .formatter.exs
in the parent directories of your application, and it stops at the app
root. If it's not found in any of the parents, you get default options.