Improve completions with default arguments
A couple of updates ago, elixir sense started returning the defaults
of arguments along with the names of the arguments. This make our
completions much worse, as the arity of the completion didn't match
the arity of the function selected *and* the defaults would get pasted
into the completion. This meant you had to remove the default from the
completion, which was beyond annoying.

These changes strip out defaults and ensure that the arity and
arguments match the completion result generated by elixir sense.