Support struct fields completions
I created an `Ast` module in the common app,
so we can use the `cursor_path/2` in both `remote_control` and `server` apps.

When completing struct fields,
we need to avoid adding a newline to the position as it may cause an AST parse error while completing the second field.

To facilitate the conversion of struct fields provided by remote_control into a complete format,
I added a translate function for the struct_fields candidates.
I also created a new/2 in the `Condidate.StructField` module.
All the desired fields are now in the struct arguments, so I set the `call?` field to false.