Dependency structs were not being detected
This is likely because we were picking up dependent structs via a
project compile, and even if forced, that won't recompile
dependencies, which means they slip by.

This change actually looks at every elixir module to decide if it has
structs or not, and tries to unload those modules that it had to load
to save resources. This negatively affects performance, but until we
have a cache of sorts, this is the best we can do (and the cache will
definitely make things tricker, as invalidation is very tricky)

Fixes #204