Support projects having the same directory name as a dependency
In order to use the `Mix.Project.in_project`, you need to give a name
for the project, and we were defaulting to the atomized name of their
project directory. The problem with this approach is if you named your
project "jason" and it depends on "jason", then your project will
conflict with the `jason` app when it comes time to run
`deps.compile`. By naming your project after the module defined in
`mix.exs`, this avoids the conflict, but still keeps the naming
consistent (mix defaults to this project name).

We can't use the project's `:app` key, because umbrellas don't define
it.

Another approach would be to give the project a unique ID when it's
created and to use that as the name of the project.