Alias module
Built an alias module with submodules that can look at a bit of source
code and determine what the aliases are at a given position.
It uses a stack of scopes, where each scope has a set of aliases
associated with it. When we reach a do block, we push another scope onto
the stack, and when we find its end, we pop the current scope from the
stack.
When we need aliases, we simply reverse the stack and add each scope's
aliases to an accumulator. This way newer aliases overwrite older ones,
and you're left with the current set of aliases.