Improved actions workflow
The old workflow was serial, and was taking way too long after the
changes to the product structure. This is due to the dialyzer plt having
to be modified by all the different usages of it. Having each subproject
and the umbrella use different PLT files fixed that problem.

I also parallelized the static analysis, dialyzer and tests so they run
more quickly.