Removed double compilation (#259)
We noticed that when we'd do a release, we would have to build twice
or the server didn't work. This always irked me, as that makes no
sense, builds should be deterministic.

What was happening is that the Jason.Encoder protocol was being
consolidated, and _this_ breaks the server. Simply removing the
consolidation (which is what the double builds did) fixes the problem.

I also reworked the mix.exs file to make it a little easier to
understand.

Now we need to research why Jason.Encoder's consolidation breaks things.