Heavy refactor of namespacing
I did not like how namespacing turned out, two tasks, one with a lot
of code that didn't have defined responsibilities, it was hard to
understand, and left the code in a bad state.

After looking at some of the files in a release, it occurred to me
that we had all we needed was to perform namepacing on those files
rather than messing up the _build directory with namespaced
beams. This path lead to here, where we rewrite things in the
release's directory and its artifacts.

The effect of this is that you can make namespacing happen much more
quickly, without the clean / build / namespace / clean process that we
had before.

Oh yeah, I parallelized writing the beams. It's super fast now.