Re-enabled multiple version support
Prior to this change, running a project in a different version of
elixir or erlang than lexical was compiled under would cause an
error.

There were two reasons for this:

 1. The release is extremely sensitive to the version of elixir /
 erlang, and would not start if they didn't match. This would happen
 if include_erts is false.
 2. Even if include_erts is true, finding the elixir executable to run
 would use the inherited environment variables from the server
 process. ASDF munges the path, and would find the version of elixir
 that was used to start lexical, not the one the project requests.

The solution is to enable `use_erts` and to undo the path munging that
asdf performs. We need to do something similar with rtx.