Added default candidate case
Elixir sense has a seemingly infinite set of maps that it returns from
completion, and if we didn't explicitly handle one, the app would
crash. This is not acceptable, so I added a default case and a log.

I also had to change the place where candidates were created so nil
candidates were omitted.