Namespacing refinements
I was not changing the directory names of the namespaced apps. This
lead to some difficulties with paths, and we had to go to lengths when
to shield paths from namespacing. Well, it turns out that if you just
commit to namespacing even more, then you don't need to make special
cases.

This seems to work, and I'm going to see if I can use namespacing to
apply to a `mix install` after it happens.