Second attempt to make struct completion more consistent
The execution steps of this new algorithm are as follows:

1. If there is only one module within three generations,
only that module will be returned, e.g.: `%Lexical.Document.Changes{}`
2. If there are more structs within three generations and the first layer of the struct has ancestor structs,
the ancestor structs and more will be mixedly returned, but not the ancestor module.
3. If there are no ancestor structs, only more will be returned.

In addition, we do not return any structs beyond two generations due to `dots_count + 1`.