Introduce `Lexical.Server.CodeIntelligence.Entity`
* Add `resolve/2`

This is a general-purpose API for resolving the code entity at a given
position. Currently it only supports resolving modules (and aliased modules).
It will have immediate use for accessing docs on hover.

* Move `definition/3`

Entity resolution and go-to-definition logic are related: the first part of
going to a definition is resolving the entity under the cursor so that you know
what to go to. While go-to-definition still currently relies on `elixir_sense`,
the plan is to eventually remove that dependency, at which point it will share
logic with or use `resolve/2`.