More precise entity resolution at cursor
This commit swaps out the use of `Sourceror.parse_string/1` with a call to
`Sourceror.string_to_quoted/2` that includes the `:static_atoms_encoder`
option. This option is used to retain line/column metadata on certain atoms
(like the individual segments of an alias) that would otherwise be discarded.

An explanatory comment has been added to `Lexical.RemoteControl.CodeMod.Ast/parse/1`
for additional context.