Extract completions env into `Lexical.Ast.Env`
Completions are currently built on top of two behaviours, `Environment` and
`Builder`. The first describes a behaviour for inspecting the contextual
environment of a given position so that you can filter completions to only
those that would make sense in that context. The second describes a behaviour
for turning those completions into something that the LSP understands.

The `Environment` behaviour and implementation is generally useful outside of
the context of completions. The completions feature implemented both behaviours
in a single `Lexical.Server.CodeIntelligence.Completion.Env`. This commit
decouples the env and the builder implementations, moving the
environment-related code (and struct) to `Lexical.Ast.Env` and putting the
builder implementation in `Lexical.Server.CodeIntelligence.Completion.Builder`.