Added version compatibility check (#332)
* Added version compatibility check

Lexical now outputs the erlang and elixir versions of the VM that was
used to create them. It now checks the build and packaging directories
for these versioning files and emits an error message if the vm starts
with beam files that won't work.

* Detected version change on call to package

Another belt to wear with the suspenders. When we build a package that
overwrites another package, if the version has changed, we delete the
old compiled code and the old package and start compilation from
scratch. This definitely prevents old beam files from ending up in the
new package.