Context-aware "use" completions (#336)
* Conext-aware "use" completions

When using `use`, we would complete any module, even those without
`__using__` macros defined.

This PR filters any suggestions inside a `use` directive to ensure
that any suggested modules (or a submodule) contain a __using__ macro
of arity 1.

One thing that cropped up is that not all modules are loaded when
queried, so I had to use `:code.all_available` to generate the module
list. This wasn't super duper fast, so I had to put a cache in front
of it, which is the large part of this PR.

Another oddity is that I couldn't just send anonymous functions as
predicates, so I had to come up with a way to implement filtering
logic across distribution, so I did a matchspec-like thingy, which was
a bit ugly, so I made some syntax to make it look like function captures.