Add elixir boot script to add archives to erl path
The start_lexical.sh launch script didn't properly handle having spaces
in the path to the package; it would fail to construct paths to the
erlang archives and add them to the erlang path. This moves this work
to an elixir script, where there is no need to manage escaping spaces
in paths.