Indexing features for modules
This commit introduces indexing for module references. It currently
will only work for files to which we have the source code, turns them
into AST and then applies one or more "extractors" on it. An extractor
runs over AST and indetifies things it's looking for, and then emits
entries, which will be used to build a searchable index.

Currently, there is one extractor implemented: Module references. This
does a fairly exhaustive search for module references, wherever
they're hiding. It then uses the aliases module to find the original
source module.

Note that most of the higher level functions are just placeholders. An
indexing strategy needs to find things inside of .exs files as well as
.ex files, and we currently have no way to index core elixir (this
will be highly dependent on how elixir was installed and if we can get
the source code to the version we're running on).