ETS based storage / Fuzzy matching
This commit has ETS based file storage and a fuzzy matching system.
I was having a lot of trouble coming up with an efficient way to perform
fuzzy matching against a corpous of subjects efficiently using the data
stores. ETS is... ets, and it appears that it's impossible to use it for
any type of detailed string matching. SQLite had similar problems. While
it had a lot more complicated and helpful utilities, it just couldn't do
what I wanted when it comes to fuzziness.

Then I thought, hey, why not just store the subjects in memory with all
the places that they appear? Then you just keep that up to date, and you
remove a lot of duplication, which improves performance a lot.

Prior to this, ETS was taking around 300ms to do simple-ish matching,
and SQLite was taking around 180ms. Fuzzy can do it in 14ms. Better yet,
building a fuzzy from scratch only takes 40ms for lexical and all its
deps. This includes doing much more interesting matching.

I think we have a winner.