Tightened up fuzzy
Fuzzy was pretty coupled naming wise to references and paths. This
changes it around a bit to use the nomenclature of keys, values an
subjects.

I also played around with fuzzy matching and found that the results
were quite poor when using all of the discovered modules. This was due
to not taking the amount of match into account and not sorting in
descending order. I added unit tests and moved things around.