Add `textDocument/hover` support for modules (#331)
* Add `Lexical.Server.CodeIntelligence.Entity` containing `resolve` and `definition`

`resolve` is a general-purpose API for resolving the code entity at a given
position. Currently it only supports resolving modules.

Entity resolution and go-to-definition logic are related: the first part of
going to a definition is resolving the entity under the cursor so that you know
what to go to. While go-to-definition still currently relies on `elixir_sense`,
the plan is to eventually remove that dependency, at which point it will share
logic with or use `resolve/2`.

* Support `textDocument/hover` for modules and aliases

Add initial support for `textDocument/hover` LSP requests, showing basic
moduledocs when a module or alias is hovered.

This includes initial work on `Lexical.RemoteControl.CodeIntelligence.Docs`,
a module for looking up all documentation for a module at runtime, including
contained functions, types, callbacks, etc.