Moved dispatch to :gen_event
As I was integrating indexing into lexical, I kept having to implement
genservers that had no other point than to receive messages and take
some action when one was received. :gen_event already does this, so I
thought it would be nice to have a simple way to implement handler
modules in addition to proceses that receive events.