Include callbacks in module hover and fix `Ast` position finding
This commit currently pulls in a to-be-merged branch of `Sourceror` which
contains fixes and changes around `Sourceror.get_range/1`. These are required
in order to accurately determine the cursor position in an AST.

An additional important change that this commit makes is to add the `unescape:
false` option when converting strings to quoted forms. This is needed to
determine whether newlines in strings were escaped literals (`\n`) or
multi-line.

Take this example:

    iex> Code.string_to_quoted!(~S|"foo\nbar"|)
    "foo\nbar"

    iex> Code.string_to_quoted!(~S|"foo
    ...> bar"|)
    "foo\nbar"

    iex> Code.string_to_quoted!(~S|"foo\nbar"|, unescape: false)
    "foo\\nbar"

    iex> Code.string_to_quoted!(~S|"foo
    ...> bar"|, unescape: false)
    "foo\nbar"

Note that, with `unescape: false`, there is a difference between the strings.