Fixed index store test failure
Depending on what was run prior, the store test suite might encounter
a previously built index, which would make the create index tests
fail. The change here deletes the index in a setup_all block, so the
test suite starts with an empty index, and keeps the on_exit handler
that deletes the index at the end of each test.