Rename module base on Indexer
We need to consider two typical scenarios:
one is renaming an exact entry,
and the other is renaming both the entry and its descendants.

I differentiate between these two scenarios based on the cursor's position.
If there are no more child modules after the cursor, for example,
in the case of `defmodule TopLevel.|Entry`, then rename the `Entry` module.

However, if there are child modules after the cursor,
then rename both the Entry module and its descendants.
For example, if the cursor is at `defmodule TopLevel.|Entry.Child`,
then `TopLevel.Entry.AnotherChild` will also be renamed.