Attempt syntax corrections when formatting fails
There are some compilation/formatting errors, like `%{foo:bar}`, that
emit syntax errors that are trivially fixed. In this case, a space needs
to be added after the colon. This commit adds a simple method for
attempting to correct the syntax error and retrying the formatting
operation, and implements a correction for the keyword error previously
mentioned.