Async indexing (#371)
* Add `time/1` IEx helper to print duration to execute a fun

* Load search index asynchronously

Depending on the computer and size of the project being indexed, it was
possible for indexing to take longer than the previous 20 second
timeout. Additionally, since indexing was blocking the store GenServer,
any calls made to the Store during a long indexing would be likely to
time out (default GenServer call timeout is 5 seconds).

This commit updates the `Store.State` API to load asynchronously so that
the store can continue to respond while indexing occurs in a task, while
the indexing itself has had its timeout changed to `:infinity`.