Revamped unit tests
Calling the setup blocks in a loop effectively queued up a bunch of
on_exit handlers, which were running in sequence and failing to
actually clean up properly.

What we actaully needed to do is move the setup blocks out of the
loops, and only call the cleanup code when the tests start and end.

I also wired up destroy for mnesia to actaully delete the directory
and the database files, so it can be called from tests.