When resolving a struct, always resolve full alias
When resolving modules, we'll exclude segments trailing the cursor so
that you can walk up a module to better understand its context. For
instance, for `My.|Nested.Module`, we resolve to `My.Nested`.

Using the same behavior for structs makes less sense. If you're hovering
a module being used as a struct, often only the full alias is actually
a struct; the parents might not be. For instance, with
`%My.|Nested.Struct{}`, resolving to `My.Nested` will not yield useful
results. Because of this, the full alias is now always resolved in a
struct context, so the above would resolve to the struct
`My.Nested.Struct`.