CubDB Backend
After the nightmare that was the mnesia backend, I figured I'd try
cubdb. It's alright, the schema took some serious thinking (and I'm
sure it can be improved), but it's reasonably fast enough for our
purposes, and the code doesn't make my eyes bleed.

I'm going to be adding more tests, but it's mostly ready to be integrated.