Detect version manager the same way in all scripts
`start_lexical.sh` and `port_wrapper.sh` were using different methods to
detect the version manager. `start_lexical.sh` was updated more
recently, so I standardized on that.

I also refactored the detection function slightly to echo the result
instead of setting a global variable, since it was only used in one
place and can be easily captured. This should also make it easier to
extract this function into a shared utility if we desire.

Finally, `start_lexical.sh` wasn't doing anything to start the program
in the background in non-detected version manager case. I don't fully
understand why this is necessary in the first place, but I added it for
consistency.