Make the operational behavior of the ancestors of structures and modules more consistent
For example, find references against the ancestors of a module or structure:

Taking `Mix.Task.Compiler.Diagnostic` as an example, the references are inconsistent when the cursor is on `%Mix.Task.|Compiler.Diagnostic{}` compared to when it is on `Mix.Task.|Compiler.Diagnostic`.

I believe that maintaining consistency between them would provide a better user experience and make it easier to extend for future features, such as renaming.