Store updates only after a quiet period
This change makes the index store update to the backend after a quiet
period. This should help improve performance and cut down on churn
caused by updates while the user is editing code.

Note: This begs the question of whether or not the ETS backend should
_also_ wait to flush its data. As of now, I think this complication
can be safely eliminated.

Fixes #427