Bugfix: Document Store needs to be globally unique
Due to all nodes being connected via distribution after the ETS store
is enabled, we can no longer register a document store globally with a
non-unique name because Document.Store needs to be associated with an
editor.
This change creates some entropy when the server starts to make the
document store's name unique, and passes that entropy to the project
node.

Fixes #425