Code Actions refactor
After observing a PR for code actions, it seemed there was much to
much busywork in creating them. This PR seeks to remedy that. Now, in
order to create a code action, you just need to create a new handler
in the `remote_control` app.

The server provider now encompasses all code actions, and handles
marshalling from and to the protocol. AST transformations now reside
in remote_control, which is where they belong.

Fixes #446