Fix: Stutter when completing inside string interpolations
When completing certain things in string interpolations, the prefix
can't be processed by the tokenizer, so `token_stream` returns the EOL
from the previous line. We used the length of the EOL as the length of
the prefix, which was often not enough, so we'd get a stutter when
applying the text edit if the prefix was longer than 2 characters.

The fix is to detect the EOL, and in that case, use Code.Fragment to
parse the prefix.

Fixes #462