Fix `undefined variables` diagnostics error for HEEx files (#463)
If we simply use the quoted compiled by the EEx engine for evaluation, it will cause some false positives, such as variables defined by the parent component.

The approach this PR adopts is to evaluate twice, once with the HEEx engine compiled and once with the EEx engine compiled. This method can make the diagnostics of `undefined variables` more accurate.