Added completions for typespecs
Prior, we were ignoring typespecs as completion candidates because
they are only relevant in a couple of places. This adds detection for
those places and only emits completion candidates if they're in a
type, typep, opaque or spec.

Additionally, while in one of those contexts, it suppresses most other
completions, because they are not relevant.

Fixes #473