Heavy Refactor: Ast.Env
The Ast.Env module was accruing a lot of functionality related to
detecting the "part of speech" of code at a given point. It was
over 600 lines, and its test was approaching 1000.

Additionally, while the tests for context detection did an admirable
job for positive cases, their negative cases (ensure that this code is
_not_ a given context) were haphadard, inconsistent or missing entirely.

This PR pulls out context detection into separate modules, each with
their own test cases. What's more, the test cases now have a large
test suite, that generates both positive _and_ negative tests for each
module.

It does this by marking each case with a path of what it's testing,
and marking the code with ranges where that context should occur. The
test case then performs *exhaustive* testing on the code, going token
by token and ensuring that the given token is either in or not in the
requested context. This PR also adds over 900 automated unit tests to
enforce detection is accurate.

The end result is a lot of edge cases were found and fixed in this PR,
and some of the code can be further improved with more confidence that
the result is correct, and there won't be regressions.