Implement the extraction of the current block and usages on the left side of the "=" sign.
All variable usages follow this format: {variable_atom, meta, nil}. Therefore, we have used the process of elimination here. If a position is not a definition, it is considered a usage. To find the definition, we need to obtain the actual block at the current position.Our previous current_block function did not accurately extract the current block or scope. This commit implements the true extraction of the current block.