Added Module Loader
Calling Code.ensure_loaded during indexing was a huge
bottleneck. Apparently, trips through the language server are fairly
costly. By implementing a read-through cache in front of the code
loader we were able to reduce indexing time to 4.4 seconds from 13
seconds.

The difference was so profound, I updated all instances of calls to
`Code.ensure_loaded?` to use the loader cache.

Unfortunately, this means that we have to make common into an
application in order to start the cache.