Heavy reafacor: Moved runtime analysis to remote_control
Analysis gets complicated, and needs to run in the same environment
where the project's modules live. Because of this, we need to move any
functionality that relies on runtime introspection to
`remote_control`. The only analysis that can be done in `common` is
analysis that only takes the AST into account.

We're going to need to do some work in the AST modules as well in
order to tease out analysis and ensure that it only happens in
`remote_control`.