Added extractor for function references, improved test cases (#508)
* Added extractor for function references, improved test cases

Added an extractor that scans the code for any function references,
and indexes them.

Also created an extractor case that has functions that help make
testing extraction of various things easier.

* Added Module Loader

Calling Code.ensure_loaded during indexing was a huge
bottleneck. Apparently, trips through the language server are fairly
costly. By implementing a read-through cache in front of the code
loader we were able to reduce indexing time to 4.4 seconds from 13
seconds.

The difference was so profound, I updated all instances of calls to
`Code.ensure_loaded?` to use the loader cache.

Unfortunately, this means that we have to make common into an
application in order to start the cache.

* Added benchmarks for analysis

Added some analysis bencharmarks. These analyze enum.ex from elixir,
which is a 4500 line file with nearly evey bit of syntax in it.

I was able to knock a half-second off the benchmarks by caching
lookups for a module's mfas

* Made imports map based

Imports were previously represented as a large mapset, but this made
modifying them linear. By making them maps, you can easily overrite or
merge a module's imports, which improves performance.

Lexical used to take 4.0 seconds to index, now it takes 3.2

* Switched to expand_alias

* Ensured common only starts in the remote control app

Also fixed failing test

* stopped functions from being imported if they start with _

* got rid of parent imports

* Removed parent aliases

* Heavy reafacor: Moved runtime analysis to remote_control

Analysis gets complicated, and needs to run in the same environment
where the project's modules live. Because of this, we need to move any
functionality that relies on runtime introspection to
`remote_control`. The only analysis that can be done in `common` is
analysis that only takes the AST into account.

We're going to need to do some work in the AST modules as well in
order to tease out analysis and ensure that it only happens in
`remote_control`.

* Moved analyzer logic to Analysis

Also Removed the Analyzer module and pushed up all the submodules to Analysis