Handle compilation exiting
It is possible for compilation to exit, and we were not handling it
properly. This was complicated by the fact that builds happen inside a
GenServer, so wrapping things in a try/rescue with an exit handler
wasn't enough.

Instead, i launched another process that's monitored and awaited by
the build genserver. Then any DOWN messages are turned into
diagnostics, and reported.

Fixes #504