Find references: Functions
Implemented find references for functions.

Initially, very few references were found, and this was because we
were deduping the index by id. Removing this call allows us to see the
entire project's references, with the notable exception of imported
macros. I believe we'll need to augment the extractor with a
compilation tracer if we wish to see those. The bad news is lexical
takes around 12 seconds to compile, but around 3 seconds to index.