Added Window.send_messages
We're going to need some sort of confirmation / messaging, so I
thought it was the appropriate time to add Window.send_message. This
allows us to send a message to the user, and optionally await some
kind of response.

To accomplish this, I had to wedge in the notion of a server sent
request, which took some doing, as we don't have a mapping from
requests to responses (we sort of do now). This required doing some
work in the proto layer, as requests didn't have the ability to parse
themselves (since they originate from the server).

I tested this in both emacs and vscode, and it seems to work