Khepri backend
Earlier in the week, I came to the realization that I was writing a
database, which isn't what this project is about. An offhand comment
from Hauleth in the elixir discord made me aware of the Khepri
database by the folks who brought us rabbitMQ, and after looking at
it, it seemed to be near ideal for our project.

This commit changes the backing store of the index to use the khepri
database, which gives us a path-like structure to store data,
journaling and snapshots, transactions, stored procedures and a whole
host of other database-y things. It's a little (a lot) slower to load
the initial data set (six seconds vs. a half second), but search time
is quick, and the interface is fantastic.